+++ crypto.h
@@ -16,6 +16,6 @@
 int to_alloc(char *text);
 byte encrypt(char *key, char *text, char *crypto, char *iv);
 byte decrypt(char *key, char *crypto, char *text, char *iv, int len);
-unsigned short crc(const char *data, size_t size);
+unsigned long crc(const char *data, size_t size);
 
 #endif /* CRYPTO_H_ */
